Integrate MXTensor cat tests and select layouts by accelerator
